+++ CHANGELOG.md
@@ -28,6 +28,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - The deprecated `WithRouteTag` in `go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp`.
   The route is already added automatically for spans.
   For metrics, the alternative is to use the `WithMetricAttributesFn` option. (#8268)
+- `instrumentation/net/http/otelhttp` The `Labeler` type and its associated functions
+  (`LabelerFromContext`, `ContextWithLabeler`, `Add`, `Get`) are deprecated and will be
+  removed in a future release. Use the `WithMetricAttributesFn` option. (#8409)
 
 ### Changed